Refactor UserPlantViewSet to filter plants by user instead of box
