+++ backend/botany/api.py
@@ -14,4 +14,4 @@ class UserPlantViewSet(viewsets.ModelViewSet):
     authentication_classes = [JWTAuthentication]
 
     def get_queryset(self):
-        return UserPlant.objects.filter(user=self.request.user)
+        return UserPlant.objects.filter(box__owner=self.request.user)
