+++ backend/botany/viewsets.py
@@ -1,8 +1,13 @@
+import re
+from queryish.rest import APIModel
+from pygbif import species  # noqa
+
 from wagtail.admin.panels import TabbedInterface, FieldPanel, InlinePanel, ObjectList
 from wagtail.snippets.views.chooser import SnippetChooserViewSet
 from wagtail.snippets.views.snippets import SnippetViewSet
+from wagtail.admin.viewsets.chooser import ChooserViewSet
 
-from ..models import UserPlant
+from .models import UserPlant
 
 
 class UserPlantChooserViewSet(SnippetChooserViewSet):
@@ -38,7 +43,7 @@ class UserPlantViewSet(SnippetViewSet):
     chooser_viewset_class = UserPlantChooserViewSet
 
     public_panels = [
-        FieldPanel("user"),
+        FieldPanel("box"),
         FieldPanel("name"),
         FieldPanel("description"),
         InlinePanel("gallery_images", label="Images"),
@@ -55,5 +60,40 @@ class UserPlantViewSet(SnippetViewSet):
     )
 
     def get_queryset(self, request):
-        queryset = UserPlant.objects.filter(user=request.user)
+        queryset = UserPlant.objects.filter(box__owner=request.user)
         return queryset
+
+
+class Species(APIModel):
+
+    def __str__(self):
+        return self.name
+
+    class Meta:
+        base_url = "https://pokeapi.co/api/v2/pokemon/"
+        detail_url = "https://pokeapi.co/api/v2/pokemon/%s/"
+        fields = ["id", "name"]
+        pagination_style = "offset-limit"
+        verbose_name_plural = "plant species"
+
+    @classmethod
+    def from_query_data(cls, data):
+        return cls(
+            id=int(re.match(r'https://pokeapi.co/api/v2/pokemon/(\d+)/', data['url']).group(1)),
+            name=data['name'],
+        )
+
+    @classmethod
+    def from_individual_data(cls, data):
+        return cls(
+            id=data['id'],
+            name=data['name'],
+        )
+
+
+class SpeciesChooserViewSet(ChooserViewSet):
+    model = Species
+    choose_one_text = "Choose a species"
+
+
+species_chooser_viewset = SpeciesChooserViewSet("species_chooser")  # noqa
