--- backend/botany/viewsets/__init__.py
@@ -1,2 +1,0 @@
-from .species import species_chooser_viewset
-from .plants import UserPlantChooserViewSet, UserPlantViewSet
